fix: widget styling update
