+++ src/containers/WidgetSidebar/index.jsx
@@ -1,39 +1,13 @@
 import React from 'react';
-import { useIntl } from '@edx/frontend-platform/i18n';
-import { Card, Hyperlink, Icon } from '@edx/paragon';
-import { ArrowForward } from '@edx/paragon/icons';
 
-import moreCoursesSVG from 'assets/more-courses-sidewidget.svg';
+import LookingForChallengeWidget from './widgets/LookingForChallengeWidget';
 
-import messages from './messages';
-import './index.scss';
-
-export const arrowIcon = (<Icon className="mx-1" src={ArrowForward} />);
-
-export const WidgetSidebar = () => {
-  const { formatMessage } = useIntl();
-  return (
-    <div className="widget-sidebar px-2 mt-5 pt-3">
-      <div className="d-flex">
-        <Card orientation="horizontal" id="looking-for-challenge-widget">
-          <Card.ImageCap
-            src={moreCoursesSVG}
-            srcAlt="course side widget"
-          />
-          <Card.Body className="m-auto pr-2">
-            <h4>
-              {formatMessage(messages.lookingForChallengePrompt)}
-            </h4>
-            <h5>
-              <Hyperlink variant="brand" destination="#" className="d-flex">
-                {formatMessage(messages.findCoursesButton, { arrow: arrowIcon })}
-              </Hyperlink>
-            </h5>
-          </Card.Body>
-        </Card>
-      </div>
+export const WidgetSidebar = () => (
+  <div className="widget-sidebar px-2 mt-5 pt-3">
+    <div className="d-flex">
+      <LookingForChallengeWidget />
     </div>
-  );
-};
+  </div>
+);
 
 export default WidgetSidebar;
