+++ src/containers/WidgetSidebar/index.test.jsx
@@ -2,6 +2,8 @@ import { shallow } from 'enzyme';
 
 import WidgetSidebar from '.';
 
+jest.mock('./widgets/LookingForChallengeWidget', () => 'LookingForChallengeWidget');
+
 describe('WidgetSidebar', () => {
   describe('snapshots', () => {
     test('default', () => {
