+++ src/containers/WidgetSidebar/widgets/LookingForChallengeWidget/index.scss
