+++ src/containers/WidgetSidebar/widgets/LookingForChallengeWidget/index.test.jsx
@@ -1,0 +1,12 @@
+import { shallow } from 'enzyme';
+
+import LookingForChallengeWidget from '.';
+
+describe('LookingForChallengeWidget', () => {
+  describe('snapshots', () => {
+    test('default', () => {
+      const wrapper = shallow(<LookingForChallengeWidget />);
+      expect(wrapper).toMatchSnapshot();
+    });
+  });
+});
